Refactor API views and utils
