+++ memegen/routes/_settings.py
@@ -1,0 +1,4 @@
+GITHUB_SLUG = "jacebrowning/memegen"
+GITHUB_BASE = "https://raw.githubusercontent.com/{}/master/".format(GITHUB_SLUG)
+CONTRIBUTING_URL = GITHUB_BASE + "CONTRIBUTING.md"
+CHANGES_URL = GITHUB_BASE + "CHANGELOG.md"
