+++ memegen/routes/api_magic.py
@@ -2,7 +2,7 @@ from flask import Blueprint, current_app as app, redirect
 
 from ..domain import Text
 
-from ._common import route
+from ._utils import route
 
 
 blueprint = Blueprint('magic', __name__)
