+++ tests/test_api_magic.py
