+++ tests/test_api_templates.py
