--- tests/test_routes_static.py
@@ -1,12 +1,0 @@
-# pylint: disable=no-self-use,misplaced-comparison-constant
-
-from .conftest import load
-
-
-class TestStatic:
-
-    def test_get_index(self, client):
-        response = client.get("/")
-
-        assert 200 == response.status_code
-        assert 'href="/api"' in load(response, as_json=False)
